Minor changes, added cpp file, and retested on unittests.
